--- app/src/main/java/com/vkochenkov/filmscatalog/data/api/pojo/DataItem.kt
@@ -1,7 +1,0 @@
-package com.vkochenkov.filmscatalog.data.api.pojo
-
-import com.google.gson.annotations.SerializedName
-
-data class DataItem(
-    @SerializedName("attributes") val attributes: Attributes
-)
