--- app/src/main/java/com/vkochenkov/filmscatalog/data/api/pojo/PosterImage.kt
@@ -1,11 +1,0 @@
-package com.vkochenkov.filmscatalog.data.api.pojo
-
-import com.google.gson.annotations.SerializedName
-
-data class PosterImage(
-    @SerializedName("tiny") val tiny: String,
-    @SerializedName("small") val small: String,
-    @SerializedName("medium") val medium: String,
-    @SerializedName("large") val large: String,
-    @SerializedName("original") val original: String
-)
