--- app/src/main/java/com/vkochenkov/filmscatalog/data/api/pojo/ResponseFromApi.kt
@@ -1,7 +1,0 @@
-package com.vkochenkov.filmscatalog.data.api.pojo
-
-import com.google.gson.annotations.SerializedName
-
-data class ResponseFromApi(
-    @SerializedName("data") val data: List<DataItem>
-)
