+++ app/src/main/java/com/vkochenkov/filmscatalog/model/DataStorage.kt
@@ -1,4 +1,4 @@
-package com.vkochenkov.filmscatalog.data
+package com.vkochenkov.filmscatalog.model
 
 //пока этот класс вместо базы данных
 object DataStorage {
