+++ app/src/main/java/com/vkochenkov/filmscatalog/model/Repository.kt
@@ -1,0 +1,4 @@
+package com.vkochenkov.filmscatalog.model
+
+class Repository {
+}
