+++ app/src/main/java/com/vkochenkov/filmscatalog/view/MainActivity.kt
@@ -1,13 +1,13 @@
-package com.vkochenkov.filmscatalog.presentation.view
+package com.vkochenkov.filmscatalog.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.vkochenkov.filmscatalog.R
-import com.vkochenkov.filmscatalog.presentation.view.dialogs.ExitDialog
-import com.vkochenkov.filmscatalog.presentation.view.fragments.AppInfoFragment
-import com.vkochenkov.filmscatalog.presentation.view.fragments.FavouriteFilmsListFragment
-import com.vkochenkov.filmscatalog.presentation.view.fragments.FilmsListFragment
+import com.vkochenkov.filmscatalog.view.dialogs.ExitDialog
+import com.vkochenkov.filmscatalog.view.fragments.AppInfoFragment
+import com.vkochenkov.filmscatalog.view.fragments.FavouriteFilmsListFragment
+import com.vkochenkov.filmscatalog.view.fragments.FilmsListFragment
 
 class MainActivity : AppCompatActivity() {
