+++ app/src/main/java/com/vkochenkov/filmscatalog/view/dialogs/ExitDialog.kt
@@ -1,4 +1,4 @@
-package com.vkochenkov.filmscatalog.presentation.view.dialogs
+package com.vkochenkov.filmscatalog.view.dialogs
 
 import android.app.Activity
 import android.app.Dialog
