+++ app/src/main/java/com/vkochenkov/filmscatalog/view/recycler/FilmViewHolder.kt
@@ -1,4 +1,4 @@
-package com.vkochenkov.filmscatalog.presentation.view.recycler
+package com.vkochenkov.filmscatalog.view.recycler
 
 import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
@@ -12,13 +12,13 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.request.target.CustomTarget
 import com.bumptech.glide.request.transition.Transition
 import com.vkochenkov.filmscatalog.R
-import com.vkochenkov.filmscatalog.data.Film
+import com.vkochenkov.filmscatalog.model.Film
 
-class FavouriteFilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    val filmTitle: TextView = itemView.findViewById(R.id.item_favourite_film_title)
-    val filmImage: ImageView = itemView.findViewById(R.id.item_favourite_film_image)
-    val filmDetailsBtn: Button = itemView.findViewById(R.id.item_favourite_film_details_btn)
-    val filmDeleteBtn: ImageView = itemView.findViewById(R.id.item_favourite_film_delete_btn)
+class FilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+    val filmTitle: TextView = itemView.findViewById(R.id.item_film_title)
+    val filmImage: ImageView = itemView.findViewById(R.id.item_film_image)
+    val filmDetailsBtn: Button = itemView.findViewById(R.id.item_film_details_btn)
+    val filmLikeBtn: ImageView = itemView.findViewById(R.id.item_film_like_btn)
 
     fun bind(item: Film) {
         filmTitle.text = item.title
@@ -27,11 +27,10 @@ class FavouriteFilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView
         Glide.with(itemView.context)
             .asBitmap()
             .load(item.imageUrl)
-            .into(object : CustomTarget<Bitmap>() {
+            .into(object : CustomTarget<Bitmap>(){
                 override fun onResourceReady(resource: Bitmap, transition: Transition<in Bitmap>?) {
                     filmImage.setImageBitmap(resource)
                 }
-
                 override fun onLoadCleared(placeholder: Drawable?) {
                     // this is called when imageView is cleared on lifecycle call or for
                     // some other reason.
@@ -45,6 +44,12 @@ class FavouriteFilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView
             filmTitle.setTextColor(ContextCompat.getColor(itemView.context, R.color.colorAccent))
         } else {
             filmTitle.setTextColor(ContextCompat.getColor(itemView.context, R.color.colorBlack))
+        }
+
+        if (item.liked) {
+            filmLikeBtn.setImageResource(R.drawable.ic_heart_fill)
+        } else {
+            filmLikeBtn.setImageResource(R.drawable.ic_heart_empty)
         }
     }
 }
